# openapi/openapi-contract-consumption-be-tier1-v1.yaml
info:
  title: Contract Consumption API
  description: |-
    This API enables the discovery, negotiation, and consumption of data contracts
    between data providers and consumers in a secure and standardized way. It supports transfer initiation,
    contract negotiation, and resource address management.

    ## Business Context

    The API facilitates data exchange in a federated data space. Consumers can search for data offers, negotiate contracts,
    and initiate data transfers from providers. The API ensures secure and traceable interactions between participants.


    ### Key Actors

    - **Provider**: Offers data assets.
    - **Consumer**: Requests and consumes data.

    ### Typical Flow

    1. Search for data offers.

    2. Negotiate a contract.

    3. Initiate a transfer.

    4. Monitor transfer status.
  contact:
    name: Simpl Programme
    url: https://simpl-programme.ec.europa.eu/
    email: cnect-simpl@ec.europa.eu
  license:
    name: European Union Public License (EUPL) 1.2
    url: https://interoperable-europe.ec.europa.eu/sites/default/files/custom-page/attachment/eupl_v1.2_en.pdf
  version: "1.0"
servers:
- url: https://contract-consumption-be.dev.simpl-europe.eu/v1
security:
- bearerAuth: []
openapi: 3.0.1
components:
  securitySchemes:
    bearerAuth:
      type: http
      description: IAA cloud gateway JWT token
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    InputObject:
      type: object
      properties:
        in:
          type: string
          enum:
          - body
          - path
          - query
          - header
        name:
          type: string
        value:
          type: object
    ContractNegotiation:
      type: object
      properties:
        contractAgreementId:
          type: string
        state:
          type: string
        counterPartyAddress:
          type: string
        counterPartyId:
          type: string
        errorDetail:
          type: string
        protocol:
          type: string
        type:
          type: string
        createdAt:
          type: integer
          format: int64
        '@id':
          type: string
    TransferProcess:
      type: object
      properties:
        finalState:
          type: string
        assetId:
          type: string
        contractId:
          type: string
        correlationId:
          type: string
        transferType:
          type: string
        errorDetail:
          type: string
        type:
          type: string
        stateTimestamp:
          type: integer
          format: int64
        '@id':
          type: string
        state:
          type: string
    JsonNode:
      type: object
    ContractNegotiationId:
      type: object
      properties:
        contractNegotiationId:
          type: string
    InputValidationIssue:
      type: object
      properties:
        type:
          type: string
          format: uri
        href:
          type: string
          format: uri
        title:
          type: string
        detail:
          type: string
        in:
          type: string
          enum:
          - body
          - path
          - query
          - header
        name:
          type: string
        value:
          type: object
        inputs:
          type: array
          items:
            $ref: '#/components/schemas/InputObject'
        additionalProperties:
          type: object
          additionalProperties:
            type: object
    ContractNegotiationRequest:
      required:
      - assetId
      - contractDefinitionId
      - providerEndpoint
      type: object
      properties:
        providerEndpoint:
          minLength: 1
          type: string
        contractDefinitionId:
          minLength: 1
          type: string
        assetId:
          minLength: 1
          type: string
...
    InternalServerErrorProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        href:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        additionalProperties:
          type: object
          additionalProperties:
            type: object
...
    BadRequestProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        href:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        issues:
          type: array
          items:
            $ref: '#/components/schemas/InputValidationIssue'
        additionalProperties:
          type: object
          additionalProperties:
            type: object
    TransferProcessId:
      type: object
      properties:
        transferProcessId:
          type: string
...
    TransferRequest:
      required:
      - contractId
      - dataDestination
      - providerEndpoint
      - templateId
      type: object
      properties:
        providerEndpoint:
          minLength: 1
          type: string
        contractId:
          minLength: 1
          type: string
        templateId:
          minLength: 1
          type: string
        dataDestination:
          $ref: '#/components/schemas/JsonNode'
paths:
  /transfers:
    post:
      tags:
      - Transfer Process
      summary: Initiate a new transfer process
      description: Starts a data transfer process using a previously negotiated contract. This operation enables the actual movement of data from provider to consumer.
      operationId: startTransfer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
            examples:
              transferRequest:
                description: transferRequest
                value:
                  providerEndpoint: https://edc-provider.dev.simpl-europe.eu/protocol
                  contractId: 4a7f28e5-33a9-4f9e-9de9-a4f687502bce
                  templateId: "1"
                  dataDestination:
                    type: IonosS3
                    region: de
                    storage: s3-eu-central-1.ionoscloud.com
                    bucketName: simpl-cons
                    objectName: european_health_data.csv
                    path: folder1/
                    keyName: test-key-name
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorProblem'
              examples:
                Internal Server Error Example:
                  description: Internal Server Error Example
                  value:
                    type: urn:problem-type:simpl:internalServerError
                    title: Internal Server Error
                    status: 500
                    detail: Unexpected internal error
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/BadRequestProblem'
              examples:
                Bad Request Error Example:
                  description: Bad Request Error Example
                  value:
                    type: urn:problem-type:simpl:validationError
                    title: Invalid payload
                    status: 400
                    detail: missing required arguments
            '*/*':
              schema:
                $ref: '#/components/schemas/BadRequestProblem'
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferProcessId'
              examples:
                successfulTransferStart:
                  description: successfulTransferStart
                  value:
                    transferProcessId: c76e83a1-2271-4143-a3f5-4f64d4cb5ecc
...
  /contracts/{contractNegotiationId}:
    get:
      tags:
      - Contract
      summary: Return the contract negotiation status
      description: Retrieves the current status of a contract negotiation. Useful for tracking the progress and outcome of the negotiation process.
      operationId: getContractNegotiationStatus
      parameters:
      - name: contractNegotiationId
        in: path
        description: The ID of the contract negotiation to retrieve the status for
        required: true
        schema:
          minLength: 1
          type: string
      responses:
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorProblem'
              examples:
                Internal Server Error Example:
                  description: Internal Server Error Example
                  value:
                    type: urn:problem-type:simpl:internalServerError
                    title: Internal Server Error
                    status: 500
                    detail: Unexpected internal error
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/BadRequestProblem'
              examples:
                Bad Request Error Example:
                  description: Bad Request Error Example
                  value:
                    type: urn:problem-type:simpl:validationError
                    title: Invalid payload
                    status: 400
                    detail: missing required arguments
            '*/*':
              schema:
                $ref: '#/components/schemas/BadRequestProblem'
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractNegotiation'
              examples:
                negotiationStatus:
                  description: negotiationStatus
                  value:
                    '@id': 9e82cdd6-4aba-4a77-8dcc-f1a9f56d82c5
                    contractAgreementId: ca-12345-abcde
                    state: CONFIRMED
                    counterPartyAddress: https://provider-edc.example.com/api/v1/dsp
                    counterPartyId: provider-id-123456
                    protocol: dataspace-protocol-http
                    type: CONSUMER
                    createdAt: 1716825600000
  /contracts:
    post:
      tags:
      - Contract
      summary: Initiate a new contract negotiation
      description: Initiates a contract negotiation between a consumer and a provider for a selected data offer. This step is required before any data transfer can occur.
      operationId: startContractNegotiation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContractNegotiationRequest'
            examples:
              transferRequest:
                description: transferRequest
                value:
                  providerEndpoint: http://endpoint.provider.com
                  contractDefinitionId: 961ead7f-d8b2-4fc6-8e94-e3efbe2127ea
                  assetId: asset-id
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorProblem'
              examples:
                Internal Server Error Example:
                  description: Internal Server Error Example
                  value:
                    type: urn:problem-type:simpl:internalServerError
                    title: Internal Server Error
                    status: 500
                    detail: Unexpected internal error
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/BadRequestProblem'
              examples:
                Bad Request Error Example:
                  description: Bad Request Error Example
                  value:
                    type: urn:problem-type:simpl:validationError
                    title: Invalid payload
                    status: 400
                    detail: missing required arguments
            '*/*':
              schema:
                $ref: '#/components/schemas/BadRequestProblem'
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractNegotiationId'
              examples:
                negotiationStarted:
                  description: negotiationStarted
                  value:
                    contractNegotiationId: 9e82cdd6-4aba-4a77-8dcc-f1a9f56d82c5
...
  /transfers/{transferProcessId}:
    get:
      tags:
      - Transfer Process
      summary: Returns the transfer process status
      description: Returns the current status of a data transfer process. Allows consumers to monitor the progress and completion of data delivery.
      operationId: getTransferStatus
      parameters:
      - name: transferProcessId
        in: path
        description: The ID of the transfer process to retrieve the status for
        required: true
        schema:
          minLength: 1
          type: string
      responses:
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorProblem'
              examples:
                Internal Server Error Example:
                  description: Internal Server Error Example
                  value:
                    type: urn:problem-type:simpl:internalServerError
                    title: Internal Server Error
                    status: 500
                    detail: Unexpected internal error
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/BadRequestProblem'
              examples:
                Bad Request Error Example:
                  description: Bad Request Error Example
                  value:
                    type: urn:problem-type:simpl:validationError
                    title: Invalid payload
                    status: 400
                    detail: missing required arguments
            '*/*':
              schema:
                $ref: '#/components/schemas/BadRequestProblem'
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferProcess'
              examples:
                transferProcessStatus:
                  description: transferProcessStatus
                  value:
                    '@id': c76e83a1-2271-4143-a3f5-4f64d4cb5ecc
                    state: COMPLETED
                    finalState: COMPLETED
                    assetId: asset-123456
                    contractId: contract-789012
                    correlationId: corr-345678
                    transferType: HttpData
                    type: CONSUMER
                    stateTimestamp: 1716825600000